Fixed resset password message templates

+++ metatierrascol/core/resetPasswordViews.py
@@ -14,8 +14,8 @@ from metatierrascol.settings import API_URL, TEMPLATE_ASSETS_URL
 from .commonlibs import emails
 from .resetPasswordSerializers import ResetPasswordRequestSerializer, ResetPasswordSerializer
 
-
-class RequestPasswordReset(generics.GenericAPIView):
+#1. The user request an email. This view sends the email
+class RequestResetPasswordEmail(generics.GenericAPIView):
     """
     Sends an email with a link to see a form to put the new password
     """
@@ -42,6 +42,23 @@ class RequestPasswordReset(generics.GenericAPIView):
         else:
             return Response({'success': ['Le hemos enviado un email para reestablecer la contraseña2']}, status=status.HTTP_200_OK)
 
+
+#2. On ckiking in the link of the email send a get request
+#   to this view that sends the reset password form
+@api_view(http_method_names=['GET'])
+@permission_classes((permissions.AllowAny,))
+def sendResetPasswordForm(request, token):
+    reset_obj = PasswordReset.objects.filter(token=token).first()  
+    if not reset_obj:
+        t=get_template(template_name='passwordreset/custom_password_reset_message.html')
+        return HttpResponse(t.render({'message':'Token Inválido', 'TEMPLATE_ASSETS_URL':TEMPLATE_ASSETS_URL}))
+ 
+    t=get_template(template_name='passwordreset/custom_password_reset_form.html')
+    endpoint_url = API_URL + f'core/perform_reset_password/{token}/'
+    return HttpResponse(t.render({'token': token, 'endpoint_url': endpoint_url, 'TEMPLATE_ASSETS_URL':TEMPLATE_ASSETS_URL}))
+
+#3. On sending the previous form, this view receives the new password
+#   and changes the password
 class PerformResetPassword(generics.GenericAPIView):
     """
     Receives the token, and the new password twice:
@@ -60,10 +77,11 @@ class PerformResetPassword(generics.GenericAPIView):
         #if new_password != confirm_password:
         #    return Response({"error": ["Las contraseñas no coinciden"]}, status=400)
         
+        t=get_template(template_name='passwordreset/custom_password_reset_message.html')
         reset_obj = PasswordReset.objects.filter(token=token).first()
         
         if not reset_obj:
-            return Response({'error':['Token inválido']}, status=400)
+            return HttpResponse(t.render({'message':'Token Inválido', 'TEMPLATE_ASSETS_URL':TEMPLATE_ASSETS_URL}))
         
         user = User.objects.filter(email=reset_obj.email).first()
                 
@@ -71,22 +89,22 @@ class PerformResetPassword(generics.GenericAPIView):
             data=request.data.copy()
             data['email']=user.email
             serializer = self.serializer_class(data=data)
-            serializer.is_valid(raise_exception=True)
+            serializer.is_valid(raise_exception=False)
+            if not serializer.is_valid():
+                if data['new_password'] != data['confirm_password']:
+                    return HttpResponse(t.render({'message':'Las contraseñas no coinciden', 'TEMPLATE_ASSETS_URL':TEMPLATE_ASSETS_URL}))
+                else:
+                    return HttpResponse(t.render({'message':'Contraseña demasiado débil, demasiadoparecida a su email, o demasiado común. Introduzca mayúsculas, números y símbolos', 'TEMPLATE_ASSETS_URL':TEMPLATE_ASSETS_URL}))
+
             data = serializer.validated_data
             user.set_password(data['new_password'])
             user.save()
             reset_obj.delete()
-            
-            return Response({'success':['Contraseña actualizada']})
-        else: 
-            return Response({'success':['Contraseña actualizada2']})
+
+            return HttpResponse(t.render({'message':'Contraseña cambiada con éxito','TEMPLATE_ASSETS_URL':TEMPLATE_ASSETS_URL}))
+        else:
+            #el email no existe
+            return HttpResponse(t.render({'message':'Contraseña cambiada con éxito2','TEMPLATE_ASSETS_URL':TEMPLATE_ASSETS_URL}))
 
 
 
-#@method_decorator(csrf_exempt, name='dispatch')
-@api_view(http_method_names=['GET'])
-@permission_classes((permissions.AllowAny,))
-def sendResetPasswordForm(request, token):
-    t=get_template(template_name='passwordreset/custom_password_reset_form.html')
-    endpoint_url = API_URL + f'core/reset_password/{token}/'
-    return HttpResponse(t.render({'token': token, 'endpoint_url': endpoint_url, 'TEMPLATE_ASSETS_URL':TEMPLATE_ASSETS_URL}))
